requirements.txt
pandas
python-dotenv
matplotlib
geopandas
geobr
seaborn